DCE-113: Remove Clippy Warnings
