+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "FRI_Challenger"
+name = "fri_challenger"
 version = "0.5.0"
 edition = "2024"
 
@@ -7,17 +7,17 @@ edition = "2024"
 bitflags = "2.9.4"
 rand = "0.9.2"
 num_enum = "0.7.4"
-# round = "0.1.2"
 criterion="0.7.0"
 pprof = { version = "0.15.0", features = ["flamegraph"] }
 iai-callgrind = "0.16.1"
-# bitintr="0.3.0"
-# const_for = "0.1.5" 
 clippy = "0.0.302"
-# crossbeam = "0.8.4"
 once_cell = "1.21.3"
 dhat = "0.3.3"
-arc-swap = "1.7.1"
+# bitintr="0.3.0"
+# crossbeam = "0.8.4"
+# const_for = "0.1.5" 
+# round = "0.1.2"
+
 
 [features]
 dhat-heap = []
