+++ src/engine/attacks/knight.rs
@@ -8,7 +8,7 @@ pub fn get_knight_mv(sq: usize, own: u64, _: u64, _: Color) -> u64 {
 }
 
 #[inline(always)]
-pub fn get_knight_mask(sq: usize, own: u64, _: u64, _: Color) -> u64 {
+pub fn get_knight_mask(sq: usize, _own: u64, _: u64, _: Color) -> u64 {
     KNIGHT_LOOKUP[sq]
 }
