+++ src/engine/evaluation/init_eval.rs
@@ -1,4 +1,3 @@
-use crate::engine::attacks::bishop::get_bishop_mask;
 use crate::engine::attacks::king::get_king_mask;
 use crate::engine::attacks::pawn::{
     get_all_pawn_left_att_mask, get_all_pawn_right_att_mask, get_pawn_att_mask,
@@ -7,8 +6,7 @@ use crate::engine::board::board::Board;
 use crate::engine::board::color::*;
 use crate::engine::board::piece::*;
 use crate::engine::board::square::get_file;
-use crate::engine::evaluation::common_eval::{CommonEvalTrait, KING_ATT_WEIGHT};
-use crate::engine::evaluation::eval_defs::CLR_CENTER;
+use crate::engine::evaluation::common_eval::{CLR_CENTER, CommonEvalTrait, KING_ATT_WEIGHT};
 use crate::engine::evaluation::king_eval::KingEvalTrait;
 use crate::engine::evaluation::material_eval::MaterialEvalTrait;
 use crate::engine::evaluation::mobility_eval::MobilityEvalTrait;
@@ -17,7 +15,7 @@ use crate::engine::evaluation::piece_eval::OUTPOST_RANKS;
 use crate::engine::evaluation::threats_eval::ThreatsEvalTrait;
 use crate::engine::generated::king::KING_RING;
 use crate::engine::generated::pawn::{FORWARD_SPANS_LR, PAWN_3_BEHIND_MASKS};
-use crate::engine::misc::bitboard::{Bitboard, BitboardTrait, Iterator};
+use crate::engine::misc::bitboard::{BitboardTrait, Iterator};
 use crate::engine::misc::const_utility::FILE_BITBOARD;
 
 pub const MG_LIMIT: isize = 15258;
