+++ src/engine/evaluation/king_eval.rs
@@ -2,8 +2,7 @@ use crate::engine::board::board::Board;
 use crate::engine::board::color::{Color, ColorTrait};
 use crate::engine::board::piece::*;
 use crate::engine::board::square::{get_file, get_rank};
-use crate::engine::evaluation::common_eval::CommonEvalTrait;
-use crate::engine::evaluation::eval_defs::CLR_RANK;
+use crate::engine::evaluation::common_eval::{CLR_RANK, CommonEvalTrait};
 use crate::engine::generated::pawn::{ISOLATED_PAWN_LOOKUP, PAWN_FORWARD_SPANS};
 use crate::engine::misc::bitboard::{Bitboard, BitboardTrait};
 use crate::engine::misc::const_utility::{FILE_BITBOARD, RANK_BITBOARD};
