+++ src/engine/move_generator/perft.rs
@@ -1,9 +1,8 @@
 use super::make_move::BoardMoveTrait;
 use super::mv_gen::BoardGenMoveTrait;
 use crate::engine::board::board::Board;
-use crate::engine::board::moves::{Flag, Move};
 use crate::engine::board::fen::FenTrait;
-use std::fs::File;
+use crate::engine::board::moves::Flag;
 use std::time::Instant;
 
 pub struct Stats {
@@ -129,7 +128,6 @@ pub fn init_test_func(fen: &str, depth: usize, dispaly_stats: bool) -> Stats {
 
     stats
 }
-
 
 // It is commented Out because I can't cargo build --release my engine without going into the Dev Container
 // NOTE: To use uncomment the pprof inside Cargo.toml
@@ -449,21 +447,20 @@ mod tests {
     #[test]
     fn test_all_perft_test() {
         for full_line in &ALL_PERFT_TESTS {
-            
             let mut parts = full_line.split(" ;");
             let fen = parts.next().expect("test_line must contain at least a FEN").trim();
             for part in parts {
                 let mut depth_node = part.split(" ");
 
-                let depth = depth_node.next().unwrap().replace("D", ""); 
+                let depth = depth_node.next().unwrap().replace("D", "");
                 let depth_num = depth.parse::<usize>().unwrap();
 
-                let nodes = depth_node.next().unwrap(); 
-                let nodes_num = nodes.parse::<u64>().unwrap(); 
-                
+                let nodes = depth_node.next().unwrap();
+                let nodes_num = nodes.parse::<u64>().unwrap();
+
                 let stats = init_test_func(&fen, depth_num, false);
                 assert_eq!(stats.nodes, nodes_num);
-                println!("Fen: {:?}, depth: {:?}, nodes: {:?}", fen, depth_num, nodes_num ); 
+                println!("Fen: {:?}, depth: {:?}, nodes: {:?}", fen, depth_num, nodes_num);
             }
         }
     }
@@ -600,79 +597,63 @@ pub const ALL_PERFT_TESTS: [&str; 172] = [
     "n1n5/1Pk5/8/8/8/8/5Kp1/5N1N b - - 0 1 ;D1 24 ;D2 421 ;D3 7421 ;D4 124608 ;D5 2193768",
     "n1n5/PPPk4/8/8/8/8/4Kppp/5N1N w - - 0 1 ;D1 24 ;D2 496 ;D3 9483 ;D4 182838 ;D5 3605103",
     "n1n5/PPPk4/8/8/8/8/4Kppp/5N1N b - - 0 1 ;D1 24 ;D2 496 ;D3 9483 ;D4 182838 ;D5 3605103",
-
     // // Specials
-	"3k4/3p4/8/K1P4r/8/8/8/8 b - - 0 1 ;D6 1134888",
-	"r3k2r/1b4bq/8/8/8/8/7B/R3K2R w KQkq - 0 1 ;D4 1274206",
-	"8/8/8/8/k1p4R/8/3P4/3K4 w - - 0 1 ;D6 1134888",
-	"8/8/1k6/2b5/2pP4/8/5K2/8 b - d3 0 1 ;D6 1440467",
-	"8/5k2/8/2Pp4/2B5/1K6/8/8 w - d6 0 1 ;D6 1440467",
-	"8/8/4k3/8/2p5/8/B2P2K1/8 w - - 0 1 ;D6 1015133",
-	"8/b2p2k1/8/2P5/8/4K3/8/8 b - - 0 1 ;D6 1015133",
-	"5k2/8/8/8/8/8/8/4K2R w K - 0 1 ;D6 661072",
-	"4k2r/8/8/8/8/8/8/5K2 b k - 0 1 ;D6 661072",
-	"3k4/8/8/8/8/8/8/R3K3 w Q - 0 1 ;D6 803711",
-	"r3k3/8/8/8/8/8/8/3K4 b q - 0 1 ;D6 803711",
-
-    // // en passant capture checks opponent
-	"8/8/1k6/2b5/2pP4/8/5K2/8 b - d3 0 1 ;D6 1440467",
-	"8/5k2/8/2Pp4/2B5/1K6/8/8 w - d6 0 1 ;D6 1440467",
-
-    // // avoid illegal ep(thanks to Steve Maughan)
-	"3k4/3p4/8/K1P4r/8/8/8/8 b - - 0 1 ;D6 1134888",
-	"8/8/8/8/k1p4R/8/3P4/3K4 w - - 0 1 ;D6 1134888",
-    
-    // // avoid illegal ep #2
-	"8/8/4k3/8/2p5/8/B2P2K1/8 w - - 0 1 ;D6 1015133",
+    "3k4/3p4/8/K1P4r/8/8/8/8 b - - 0 1 ;D6 1134888",
+    "r3k2r/1b4bq/8/8/8/8/7B/R3K2R w KQkq - 0 1 ;D4 1274206",
+    "8/8/8/8/k1p4R/8/3P4/3K4 w - - 0 1 ;D6 1134888",
+    "8/8/1k6/2b5/2pP4/8/5K2/8 b - d3 0 1 ;D6 1440467",
+    "8/5k2/8/2Pp4/2B5/1K6/8/8 w - d6 0 1 ;D6 1440467",
+    "8/8/4k3/8/2p5/8/B2P2K1/8 w - - 0 1 ;D6 1015133",
     "8/b2p2k1/8/2P5/8/4K3/8/8 b - - 0 1 ;D6 1015133",
-    
-	// // short castling gives check
-	"5k2/8/8/8/8/8/8/4K2R w K - 0 1 ;D6 661072",
+    "5k2/8/8/8/8/8/8/4K2R w K - 0 1 ;D6 661072",
     "4k2r/8/8/8/8/8/8/5K2 b k - 0 1 ;D6 661072",
-    
-	// // long castling gives check
-	"3k4/8/8/8/8/8/8/R3K3 w Q - 0 1 ;D6 803711",
-	"r3k3/8/8/8/8/8/8/3K4 b q - 0 1 ;D6 803711",
-    
+    "3k4/8/8/8/8/8/8/R3K3 w Q - 0 1 ;D6 803711",
+    "r3k3/8/8/8/8/8/8/3K4 b q - 0 1 ;D6 803711",
+    // // en passant capture checks opponent
+    "8/8/1k6/2b5/2pP4/8/5K2/8 b - d3 0 1 ;D6 1440467",
+    "8/5k2/8/2Pp4/2B5/1K6/8/8 w - d6 0 1 ;D6 1440467",
+    // // avoid illegal ep(thanks to Steve Maughan)
+    "3k4/3p4/8/K1P4r/8/8/8/8 b - - 0 1 ;D6 1134888",
+    "8/8/8/8/k1p4R/8/3P4/3K4 w - - 0 1 ;D6 1134888",
+    // // avoid illegal ep #2
+    "8/8/4k3/8/2p5/8/B2P2K1/8 w - - 0 1 ;D6 1015133",
+    "8/b2p2k1/8/2P5/8/4K3/8/8 b - - 0 1 ;D6 1015133",
+    // // short castling gives check
+    "5k2/8/8/8/8/8/8/4K2R w K - 0 1 ;D6 661072",
+    "4k2r/8/8/8/8/8/8/5K2 b k - 0 1 ;D6 661072",
+    // // long castling gives check
+    "3k4/8/8/8/8/8/8/R3K3 w Q - 0 1 ;D6 803711",
+    "r3k3/8/8/8/8/8/8/3K4 b q - 0 1 ;D6 803711",
     // // castling(including losing cr due to rook capture)
-	"r3k2r/1b4bq/8/8/8/8/7B/R3K2R w KQkq - 0 1 ;D4 1274206",
-	"r3k2r/7b/8/8/8/8/1B4BQ/R3K2R b KQkq - 0 1 ;D4 1274206",
-
+    "r3k2r/1b4bq/8/8/8/8/7B/R3K2R w KQkq - 0 1 ;D4 1274206",
+    "r3k2r/7b/8/8/8/8/1B4BQ/R3K2R b KQkq - 0 1 ;D4 1274206",
     // // castling prevented
-	"r3k2r/8/3Q4/8/8/5q2/8/R3K2R b KQkq - 0 1 ;D4 1720476",
-	"r3k2r/8/5Q2/8/8/3q4/8/R3K2R w KQkq - 0 1 ;D4 1720476",
-	// //  promote out of check
-	"2K2r2/4P3/8/8/8/8/8/3k4 w - - 0 1 ;D6 3821001",
-	"3K4/8/8/8/8/8/4p3/2k2R2 b - - 0 1 ;D6 3821001",
-
+    "r3k2r/8/3Q4/8/8/5q2/8/R3K2R b KQkq - 0 1 ;D4 1720476",
+    "r3k2r/8/5Q2/8/8/3q4/8/R3K2R w KQkq - 0 1 ;D4 1720476",
+    // //  promote out of check
+    "2K2r2/4P3/8/8/8/8/8/3k4 w - - 0 1 ;D6 3821001",
+    "3K4/8/8/8/8/8/4p3/2k2R2 b - - 0 1 ;D6 3821001",
     // // "# discovered check
-	"8/8/1P2K3/8/2n5/1q6/8/5k2 b - - 0 1 ;D5 1004658",
-	"5K2/8/1Q6/2N5/8/1p2k3/8/8 w - - 0 1 ;D5 1004658",
-
+    "8/8/1P2K3/8/2n5/1q6/8/5k2 b - - 0 1 ;D5 1004658",
+    "5K2/8/1Q6/2N5/8/1p2k3/8/8 w - - 0 1 ;D5 1004658",
     // // "# promote to give check
-	"4k3/1P6/8/8/8/8/K7/8 w - - 0 1 ;D6 217342",
-	"8/k7/8/8/8/8/1p6/4K3 b - - 0 1 ;D6 217342",
-
+    "4k3/1P6/8/8/8/8/K7/8 w - - 0 1 ;D6 217342",
+    "8/k7/8/8/8/8/1p6/4K3 b - - 0 1 ;D6 217342",
     // // "# underpromote to check
-	"8/P1k5/K7/8/8/8/8/8 w - - 0 1 ;D6 92683",
-	"8/8/8/8/8/k7/p1K5/8 b - - 0 1 ;D6 92683",
-
+    "8/P1k5/K7/8/8/8/8/8 w - - 0 1 ;D6 92683",
+    "8/8/8/8/8/k7/p1K5/8 b - - 0 1 ;D6 92683",
     // // "# self stalemate
-	"K1k5/8/P7/8/8/8/8/8 w - - 0 1 ;D6 2217",
-	"8/8/8/8/8/p7/8/k1K5 b - - 0 1 ;D6 2217",
-
+    "K1k5/8/P7/8/8/8/8/8 w - - 0 1 ;D6 2217",
+    "8/8/8/8/8/p7/8/k1K5 b - - 0 1 ;D6 2217",
     // // stalemate/checkmate:
-	"8/k1P5/8/1K6/8/8/8/8 w - - 0 1 ;D7 567584",
-	"8/8/8/8/1k6/8/K1p5/8 b - - 0 1 ;D7 567584",
-
+    "8/k1P5/8/1K6/8/8/8/8 w - - 0 1 ;D7 567584",
+    "8/8/8/8/1k6/8/K1p5/8 b - - 0 1 ;D7 567584",
     // // double check
-	"8/8/2k5/5q2/5n2/8/5K2/8 b - - 0 1 ;D4 23527",
-
+    "8/8/2k5/5q2/5n2/8/5K2/8 b - - 0 1 ;D4 23527",
     // // short castling impossible although the rook never moved away from its corner
-	"1k6/1b6/8/8/7R/8/8/4K2R b K - 0 1 ;D5 1063513",
-	"4k2r/8/8/7r/8/8/1B6/1K6 w k - 0 1 ;D5 1063513",
-
+    "1k6/1b6/8/8/7R/8/8/4K2R b K - 0 1 ;D5 1063513",
+    "4k2r/8/8/7r/8/8/1B6/1K6 w k - 0 1 ;D5 1063513",
     // // long castling impossible although the rook never moved away from its corner
-	"1k6/8/8/8/R7/1n6/8/R3K3 b Q - 0 1 ;D5 346695",
+    "1k6/8/8/8/R7/1n6/8/R3K3 b Q - 0 1 ;D5 346695",
     "r3k3/8/1N6/r7/8/8/8/1K6 w q - 0 1 ;D5 346695",
 ];
